Fri, Oct 14, 2022, 6:59 PM +00:00

+++ pset4/filter/helpers.c
@@ -92,9 +92,9 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
                 }
             }
 
-            image[i][j].rgbtRed = round(somaRed / count);
-            image[i][j].rgbtGreen = round(somaGreen / count);
-            image[i][j].rgbtBlue = round(somaBlue / count);
+            image[i][j].rgbtRed = (int)round(somaRed / count);
+            image[i][j].rgbtGreen = (int)round(somaGreen / count);
+            image[i][j].rgbtBlue = (int)round(somaBlue / count);
 
             }
         }
@@ -105,6 +105,6 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
 // Detect edges
 void edges(int height, int width, RGBTRIPLE image[height][width])
 {
-    
+
     return;
 }
